D {brnn}	R Documentation

Genomic dominant relationship matrix for the Jersey dataset.

Description

This matrix was calculated by using the dominance incidence matrix derived from 33,267 Single Nucleotide Polymorphisms (SNPs) information on 297 individually cows,

D=\frac{X_d X_d'}{2 ∑_{j=1}^p (p_j^2+q_j^2) p_j q_j},

where

X_d is the design matrix for allele substitution effects for dominance.

p_j is the frecuency of the second allele at locus j and q_j=1-p_j.

Source

University of Wisconsin at Madison, USA.

[Package brnn version 0.6 Index]