G {brnn}	R Documentation

Genomic additive relationship matrix for the Jersey dataset.

Description

A matrix, similar to this was used in Gianola et al. (2011) for predicting milk, fat and protein production in Jersey cows. In this software version we do not center the incidence matrix for the additive effects.

G=\frac{X_a X_a'}{2∑_{j=1}^p p_j (1-p_j)},

where

X_a is the design matrix for allele substitution effects for additivity.

p_j is the frecuency of the second allele at locus j and q_j=1-p_j.

Source

University of Wisconsin at Madison, USA.

References

Gianola, D. Okut, H., Weigel, K. and Rosa, G. 2011. "Predicting complet quantitative traits with Bayesian neural networks: a case study with Jersey cows and wheat". BMC Genetics.

[Package brnn version 0.6 Index]