predict.brnn {brnn}	R Documentation

predict.brnn

Description

The function produces the predictions for a two-layer feed-forward neural network.

Usage

   ## S3 method for class 'brnn'
predict(object,newdata,...)

Arguments

object
an object of the class brnn as returned by brnn

newdata
matrix or data frame of test examples. A vector is considered to be a row vector comprising a single case.

...
arguments passed to or from other methods.

Details

This function is a method for the generic function predict() for class "brnn". It can be invoked by calling predict(x) for an object x of the appropriate class, or directly by calling predict.brnn(x) regardless of the class of the object.

Value

A vector containing the predictions

[Package brnn version 0.6 Index]