un_normalize {brnn}	R Documentation

un_normalize

Description

Internal function for going back to the original scale.

Usage

   un_normalize(z,base,spread)

Arguments

z
a vector or matrix with values normalized between -1 and 1, this vector was obtained when normalizing a vector or matrix x.

base
If z is a vector, base is the minimum of x. If x is a matrix, base is a vector with the minimum for each of the columns of the matrix x.

spread
if z is a vector, spread=base-max(x). If x is a matrix, spread is a vector calculated for each of the columns of x.

Details

x=base+0.5*spread*(z+1)

Value

A vector or matrix with the resulting un normalized values.

[Package brnn version 0.6 Index]